+++ rott/rt_stat.h
@@ -151,7 +151,7 @@ typedef struct statstruct
 		  short int         whichstat;
 		  short int         areanumber;
 
-		  long              linked_to;
+		  intptr_t          linked_to;
 		  struct statstruct *statnext;
 		  struct statstruct *statprev;
 		  struct statstruct *nextactive;
@@ -167,7 +167,7 @@ typedef struct respstruct
 		  int               spawnz;
 		  struct respstruct *next;
 		  struct respstruct *prev;
-                  long              linked_to;
+		  intptr_t          linked_to;
 } respawn_t;
 
 
@@ -204,8 +204,8 @@ extern   animwall_t     animwalls[MAXANIMWALLS];
 void Set_NewZ_to_MapValue(fixed*,int,const char*,int,int);
 void RemoveFromFreeStaticList(statobj_t*);
 void CheckCriticalStatics(void);
-void ActivateLight(long);
-void DeactivateLight(long);
+void ActivateLight(intptr_t);
+void DeactivateLight(intptr_t);
 void TurnOnLight(int,int);
 void TurnOffLight(int,int);
 void MakeStatActive(statobj_t*);
